XW-1779: subtitle trait
